requirements.txt
requests>=2.28.0
httpx>=0.27.0
pyotp>=2.9.0
pytest-httpx>=0.35.0
pyTelegramBotAPI>=4.29.1
anthropic>=0.90.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
websocket-client>=1.6.0
websockets>=12.0
pyarrow>=14.0.0
pydantic>=2.0.0

scipy>=1.11.0

psutil==7.2.2
PyYAML>=6.0
